Splitting src/log/levels.hpp itno hpp/cpp
